AI/requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
databricks-cli==0.17.8
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
docker==6.1.3
entrypoints==0.4
evaluate==0.4.0
executing==2.0.0
filelock==3.12.4
Flask==2.3.3
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mlflow==2.7.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipenv==2023.9.8
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
regex==2023.8.8
requests==2.31.0
responses==0.18.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.33.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
virtualenv==20.24.5
wcwidth==0.2.8
websocket-client==1.6.3
Werkzeug==3.0.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
